v1.2.0:
Issue #4: LinearProgressionFacade - no validation whether the schema's applied versions are included in the user-supplied list of versions

Validations were put in place on all public API methods except for isVersions(). This includes the class constructor. Therefore, the constructor
now may throw a LiquibaseException if there is an error while trying to read the database metadata information. This minor change in the API is what triggered the bump in version from 1.1.1 -> 1.2.0.

This extra validation slows the implementation down substantially, however correctness and consistency is valued a lot more than performance in this API.
Task #4 - LinearProgressionFacade - no validation whether the schema's applied versions are included in the user-supplied list of versions